:recycle: (PC-5504) move PlaceCaption to a specific atom
